backend/requirements.txt
django
djangorestframework
djangorestframework-jwt
django-rest-knox
django-cors-headers
django-cors-middleware
django-extensions
django-uuidfield
drf-yasg
pylint
pylint-django
isort
Pillow
